Added objects and scripts. Modified scripts aswell.
I modified the movement scripts. I made changes to the head movement, so
that when sprinting, less movement will be made.
I also made changes to the mouse sensitivity when sprinting. Now, the
sensitivity is almost equal to the sensitivity when not sprinting.

I added a crouching script which now allows the player to crouch. The
crouching will be activated by pressing the c key on the keyboard. I
also added an object which can only be passed if the player crouches
underneath it.
I also added an onscreen prompt for the user, explaining how to crouch.
To do this, I had to modify several scripts and also add a promptpoint,
which activates the prompt.