Added bugs which wanders randomly around
Added an object and textured it so it looks like a caterpillar. Added a
script which controls the movement of the bug. Right now it just wanders
randomly around, turning around when it hits something. There are some
bugs with the movement of the bugs: sometimes they get stuck in a wall,
or they will be lifted off ground if hitting a wall at a certain angle.
I will look into this later, and try to correct.

The bugs have been scattered around in Level 1. I will add some to level
2 as well.