Added moving spiders and a barrel object
Added a spider object, and a moveSpider script, which moves the spider
around on the wall it is connected to. Right now the script only works
when the rotation of the spider is equal to x = 0, y = 90. Look at the
other spiders in level 1 if you want to add some more. I'm planning to
expand the script so that the spiders can be rotated in any direction in
the y axis, so that they can be placed on any walls.

I scattered some spiders around in Level 1.

I also added a barrel object, but I haven't placed any of them in any
levels yet.