Implemented basic functionality of 2nd puzzle
Added MoveWalls script, which moves the two walls in a corridor move
towards eachother. The player then has a certain time to run through the
corridor.

The script adds the basic functionality. The walls move towards
eachother, and the player dies if it touches one of the walls. Need to:

- Make the walls stop moving when they hit eachother
- Find a way to deactivate the puzzle when player dies
- Add more objects in the corridor so that the player will have to
crouch and jump to get through it